♻️ refactor: refactor the backend code for better organization (#4287)
* ♻️ refactor: move AgentRuntime

* ♻️ refactor: move errorResponse

* ♻️ refactor: move auth middleware

* ♻️ refactor: move createBizOpenAI

* 🐛 fix: fix rewriteQuery

* fix
