+++ src/app/(backend)/_deprecated/createBizOpenAI/auth.ts
