+++ src/libs/trpc/middleware/keyVaults.ts
@@ -1,6 +1,6 @@
 import { TRPCError } from '@trpc/server';
 
-import { getJWTPayload } from '@/app/(backend)/api/middleware/auth/utils';
+import { getJWTPayload } from '@/app/(backend)/middleware/auth/utils';
 import { trpc } from '@/libs/trpc/init';
 
 export const keyVaults = trpc.middleware(async (opts) => {
